change user id integer to bigint

+++ data/migrations/20201101142037_create-usersettings-table.js
@@ -1,6 +1,6 @@
 exports.up = async function (knex) {
     return await knex.schema.createTable('user_settings', function (table) {
-        table.increments();
+        table.bigIncrements();
         table.string('name');
         table.string('email');
         table.string('street');
@@ -8,7 +8,7 @@ exports.up = async function (knex) {
         table.string('zip');
         table.string('phone');
         table
-            .integer('user_id')
+            .bigInteger('user_id')
             .unsigned()
             .notNullable()
             .references('id')
